RMI Service Methods
New EntityHandler take responsibility for the EntityManager from the
RequestHandler.
New InvocationHandler resolves rmi calls.
Reorder Cover and CustomGraph methods for better readability.
Some changes by Eclipse CleanUp

+++ src/main/i5/las2peer/services/ocd/algorithms/LinkCommunitiesAlgorithm.java
@@ -232,7 +232,7 @@ private Matrix calculateEdgeSimilarities(CustomGraph graph, List<Vector> linkage
 				edge = edges.edge();
 				neighbor = edges.edge().opposite(node);
 				linkageDegree = degreeVector.get(neighbor.index());
-				linkageDegree += graph.getEdgeWeight(edge) / (double)node.degree();
+				linkageDegree += graph.getEdgeWeight(edge) / node.degree();
 				degreeVector.set(neighbor.index(), linkageDegree);
 				edges.next();
 			}
