+++ src/main/i5/las2peer/services/ocd/algorithms/SskAlgorithm.java
@@ -240,7 +240,7 @@ public class SskAlgorithm implements OcdAlgorithm {
 			}
 			else {
 				for(int i=0; i<memberships.columns(); i++) {
-					memberships.set(node.index(), i, 1d / (double)communityCount);
+					memberships.set(node.index(), i, 1d / communityCount);
 				}
 			}
 			nodes.next();
