+++ src/main/i5/las2peer/services/ocd/graphs/CustomNodeId.java
@@ -39,7 +39,7 @@ public class CustomNodeId {
  
     @Override
     public int hashCode() {
-        return (int)(id + graph.hashCode());
+        return id + graph.hashCode();
     }
 	
 }
